requirements.txt
pyTelegramBotAPI>=3.6.7
covid>=2.2.6
redis>=3.4.1
rq>=1.3.0
python-dotenv>=0.13.0